find food... i need to not crash into Walls, need a better function for that
